Move selectors into the main jsonapi file to be closer to a redux duck
